requirements.txt
Django==4.2.*
djangorestframework==3.16.*
django-extensions==4.1.*
gunicorn==23.0.*
django-environ==0.11.*
psycopg[binary]==3.2.10
django-cors-headers==4.9.0
ipython==8.37.0
yagmail==0.15.293